Fix _serialize_value duplication and unknown status handling in Agent Engine
- Move _serialize_value from operators and triggers into the hook module
  so both import from a single source of truth
- Warn on unknown query job status values in wait_for_query_agent_engine_job
  to surface unexpected API responses instead of silently looping
- Add round-trip serialize test for AgentEngineQueryJobTrigger with a
  Pydantic model config to verify _serialize_value is applied on serialize()
